+++ jdaviz/configs/default/plugins/markers/tests/test_markers_plugin.py
@@ -55,8 +55,12 @@ def test_markers_cubeviz(tmp_path, cubeviz_helper, spectrum1d_cube):
                                                       'spectral_axis:unit': 'm',
                                                       'world': (204.99988776727642,
                                                                 27.000099999955538),
+                                                      'world_ra': 204.99988776727642,
+                                                      'world_dec': 27.000099999955538,
                                                       'world:unreliable': False,
                                                       'pixel': (0, 0),
+                                                      'pixel_x': 0,
+                                                      'pixel_y': 0,
                                                       'pixel:unreliable': False,
                                                       'value': 8.0,
                                                       'value:unit': 'Jy',
@@ -139,6 +143,13 @@ def test_markers_cubeviz(tmp_path, cubeviz_helper, spectrum1d_cube):
     exp.export()
     assert os.path.isfile(filename)
 
+    # Also exports to CSV
+    filename_2 = str(tmp_path / "cubeviz_export.csv")
+    exp.plugin_table_format.selected = 'csv'
+    exp.filename.value = filename_2
+    exp.export()
+    assert os.path.isfile(filename_2)
+
     # clearing table clears markers
     mp.clear_table()
     assert mp.export_table() is None
@@ -171,6 +182,8 @@ class TestImvizMultiLayer(BaseImviz_WCS_NoWCS):
                                              'axes_y:unit': 'pix',
                                              'data_label': 'no_wcs[SCI,1]',
                                              'pixel': (5.0, 5.0),
+                                             'pixel_x': 5.0,
+                                             'pixel_y': 5.0,
                                              'pixel:unreliable': False,
                                              'value': 55.0,
                                              'value:unit': '',
@@ -195,6 +208,8 @@ class TestImvizMultiLayer(BaseImviz_WCS_NoWCS):
                                              'axes_y:unit': 'pix',
                                              'data_label': '',
                                              'pixel': (5.0, 5.0),
+                                             'pixel_x': 5.0,
+                                             'pixel_y': 5.0,
                                              'pixel:unreliable': False}
 
         mp._obj._on_viewer_key_event(self.viewer, {'event': 'keydown',
@@ -217,8 +232,12 @@ class TestImvizMultiLayer(BaseImviz_WCS_NoWCS):
                                                           'data_label': 'has_wcs[SCI,1]',
                                                           'world': (337.5187947653852,
                                                                     -20.831944164705973),
+                                                          'world_ra': 337.5187947653852,
+                                                          'world_dec': -20.831944164705973,
                                                           'world:unreliable': False,
                                                           'pixel': (5.0, 5.0),
+                                                          'pixel_x': 5.0,
+                                                          'pixel_y': 5.0,
                                                           'pixel:unreliable': False,
                                                           'value': 55.0,
                                                           'value:unit': '',